Change POST to js
Change fileDownload to use
'/documents/generate_and_send/generate_and_send.js'. Change update in
routes to use 'post' (not patch which is what happened when we used edit
to put the form on the right side). This cause us to always hit the
update method and not the generate_and_send method.

Add tests. Add id to trash can icon so we can click it in test.